+++ src/main/java/de/samply/exporter/ExporterController.java
@@ -16,6 +16,7 @@ import de.samply.exporter.response.entity.CreateQueryResponseEntity;
 import de.samply.exporter.response.entity.RequestResponseEntity;
 import de.samply.logger.BufferedLoggerFactory;
 import de.samply.logger.Logger;
+import de.samply.merger.FilesMergerManager;
 import de.samply.template.ConverterTemplateManager;
 import de.samply.utils.ProjectVersion;
 import de.samply.zip.Zipper;
@@ -34,10 +35,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import reactor.core.publisher.Flux;
 
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
+import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.time.Instant;
@@ -63,7 +61,8 @@ public class ExporterController {
     private final String httpServletRequestScheme;
     private final ExporterDbService exporterDbService;
     private final ConverterTemplateManager converterTemplateManager;
-    private Zipper zipper;
+    private final Zipper zipper;
+    private final FilesMergerManager filesMergerManager;
 
     public ExporterController(@Value(ExporterConst.HTTP_RELATIVE_PATH_SV) String httpRelativePath,
                               @Value(ExporterConst.HTTP_SERVLET_REQUEST_SCHEME_SV) String httpServletRequestScheme,
@@ -71,7 +70,8 @@ public class ExporterController {
                               ExporterDbService exporterDbService,
                               ConverterTemplateManager converterTemplateManager,
                               Zipper zipper,
-                              ExplorerManager explorerManager) {
+                              ExplorerManager explorerManager,
+                              FilesMergerManager filesMergerManager) {
         this.httpRelativePath = httpRelativePath;
         this.httpServletRequestScheme = httpServletRequestScheme;
         this.exporterCore = exporterCore;
@@ -79,6 +79,7 @@ public class ExporterController {
         this.zipper = zipper;
         this.explorerManager = explorerManager;
         this.converterTemplateManager = converterTemplateManager;
+        this.filesMergerManager = filesMergerManager;
     }
 
     @CrossOrigin(origins = "${CROSS_ORIGINS}", allowedHeaders = {"Authorization"})
@@ -351,7 +352,8 @@ public class ExporterController {
             @RequestParam(name = ExporterConst.FILE_COLUMN_PIVOT, required = false) String fileColumnPivot,
             @RequestParam(name = ExporterConst.PAGE_COUNTER, required = false) Integer pageCounter,
             @RequestParam(name = ExporterConst.PAGE_SIZE, required = false) Integer pageSize,
-            @RequestParam(name = ExporterConst.PIVOT_VALUE, required = false) String pivotValue) {
+            @RequestParam(name = ExporterConst.PIVOT_VALUE, required = false) String pivotValue,
+            @RequestParam(name = ExporterConst.MERGE_FILES, required = false) Boolean mergeFiles) {
         Optional<QueryExecution> queryExecution = exporterDbService.fetchQueryExecution(
                 queryExecutionId);
         if (queryExecution.isPresent()) {
@@ -366,9 +368,10 @@ public class ExporterController {
                 }
                 case OK -> {
                     try {
-                        yield fetchQueryExecutionFilesAndZipIfNecessary(queryExecutionId, fileFilter, fileColumnPivot, pageCounter, pivotValue, pageSize);
-                    } catch (ExporterControllerException | ZipperException | FileNotFoundException | ExplorerException |
-                             PivotIdentifierException e) {
+                        yield fetchQueryExecutionFilesAndZipOrMergeIfNecessary(
+                                queryExecutionId, fileFilter, fileColumnPivot, pageCounter, pivotValue, pageSize, mergeFiles);
+                    } catch (ExporterControllerException | ZipperException | ExplorerException |
+                             PivotIdentifierException | IOException e) {
                         yield createInternalServerError(e);
                     }
                 }
@@ -384,9 +387,10 @@ public class ExporterController {
                 new ByteArrayInputStream(queryExecutionError.getError().getBytes(StandardCharsets.UTF_8)));
     }
 
-    private ResponseEntity<InputStreamResource> fetchQueryExecutionFilesAndZipIfNecessary(
-            Long queryExecutionFileId, String fileFilter, String fileColumnPivot, Integer pageCounter, String pivotValue, Integer pageSize)
-            throws ExporterControllerException, ZipperException, FileNotFoundException, ExplorerException, PivotIdentifierException {
+    private ResponseEntity<InputStreamResource> fetchQueryExecutionFilesAndZipOrMergeIfNecessary(
+            Long queryExecutionFileId, String fileFilter, String fileColumnPivot, Integer pageCounter,
+            String pivotValue, Integer pageSize, Boolean mergeFiles)
+            throws ExporterControllerException, ZipperException, IOException, ExplorerException, PivotIdentifierException {
         List<QueryExecutionFile> queryExecutionFiles = exporterDbService.fetchQueryExecutionFilesByQueryExecutionId(
                 queryExecutionFileId);
         List<Path> files = convertToPath(queryExecutionFiles);
@@ -399,9 +403,10 @@ public class ExporterController {
         files = filterFiles(files, fileFilter);
         if (files.size() > 0) {
             if (files.size() == 1) {
-                return createResponseEntity(
-                        new InputStreamResource(new FileInputStream(files.get(0).toAbsolutePath().toString())),
-                        fetchFilename(files.get(0).getFileName().toString()), httpHeaders);
+                return createResponseEntity(files.get(0), httpHeaders);
+            } else if (mergeFiles != null && mergeFiles == true && filesMergerManager.isFileExtensionSupported(files)) {
+                Optional<Path> mergedFile = filesMergerManager.merge(files);
+                return (mergedFile.isPresent()) ? createResponseEntity(mergedFile.get(), httpHeaders) : ResponseEntity.notFound().build();
             } else {
                 Pair<InputStreamResource, String> inputStreamResourceFilenamePair = zipper.zipFiles(files);
                 return createResponseEntity(inputStreamResourceFilenamePair.getFirst(),
@@ -410,6 +415,11 @@ public class ExporterController {
         } else {
             return ResponseEntity.notFound().build();
         }
+    }
+
+    private ResponseEntity<InputStreamResource> createResponseEntity(Path path, HttpHeaders httpHeaders) throws FileNotFoundException {
+        return createResponseEntity(new InputStreamResource(new FileInputStream(path.toAbsolutePath().toString())),
+                fetchFilename(path.getFileName().toString()), httpHeaders);
     }
 
     private List<Path> convertToPath(List<QueryExecutionFile> queryExecutionFiles) {